requirements.txt
pydub
wave==0.0.2
datetime
python-dotenv==1.0.0
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
openai==0.28.0
SpeechRecognition==3.10.0
python-decouple==3.8
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
